Added ordering of contacts

+++ src/main/java/com/dect/website/dtos/KeyContactOrderDTO.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class KeyContactOrderDTO {
     private Long id;
-    private int order;
+    private Long order;
 }
